Refactor API and recommendation tests into modular packages
