+++ tests/api/test_auth.py
@@ -1,0 +1,4 @@
+"""Authentication API tests."""
+
+# Authentication endpoints are currently covered elsewhere.
+# This module exists to host future authentication scenarios.
